Backed out the AccountService interface change
